third_party/blink/web_tests/media/media-fragments/TC0053-expected.txt


Title: TC0053
Fragment: 't=foo:7&t=npt:3'
Comment: After processing name-value pairs, this appears to be equivalent to #t=3.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '3') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK
END OF TEST
